app/agents/website-read-rag/requirements.txt
# Core LangChain dependencies
langchain-core>=0.1.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.1.0
langchain-text-splitters>=0.1.0

# LangGraph for workflow management
langgraph>=0.1.0

# Vector database
chromadb>=0.4.0

# Web scraping and processing
requests>=2.31.0
beautifulsoup4>=4.12.0

# Environment and configuration
python-dotenv>=1.0.0

# Optional: For enhanced text processing
lxml>=4.9.0
html5lib>=1.1

# Optional: For async processing (if needed)
aiohttp>=3.8.0
asyncio-throttle>=1.0.0

# Optional: For caching (if needed)
redis>=4.5.0

# Optional: For logging (if needed)
loguru>=0.7.0 